REMOVE:不要なビューファイル削除
