--- src/views/About.vue
@@ -1,22 +1,0 @@
-<template>
-
-  <div>
-
-    <SectionHeader title="自己紹介" />
-
-  </div>
-
-</template>
-
-<script>
-import SectionHeader from '@/components/atoms/SectionHeader'
-
-export default {
-  components: {
-    SectionHeader
-  },
-}
-</script>
-
-<style scoped>
-</style>
