--- src/views/Experience.vue
@@ -1,22 +1,0 @@
-<template>
-
-  <div>
-
-    <SectionHeader title="業務経歴" />
-
-  </div>
-
-</template>
-
-<script>
-import SectionHeader from '@/components/atoms/SectionHeader'
-
-export default {
-  components: {
-    SectionHeader
-  },
-}
-</script>
-
-<style scoped>
-</style>
